Renamed recipes
moved grid-based and image-based under pointintime. See #141
