+++ honeybee/radiance/recipe/solaraccess/__init__.py
@@ -1,0 +1,1 @@
+"""Point-in-time recipes."""
